refactored spellItem from class to function that uses custom toggle hooks to manage state
