+++ src/components/SpellList/SpellList.js
@@ -1,12 +1,13 @@
 import React from "react";
-import SpellListItem from "./SpellListItem";
+// import SpellListItem from "./SpellListItem";
+import SpellItem from "./SpellItem";
 
 const SpellList = props => {
   return (
     <div className="spell-list">
       {Object.keys(props.spells).map((obj, i) => {
         return (
-          <SpellListItem
+          <SpellItem
             key={props.spells[i].index}
             spell={props.spells[i]}
             onSpellSelect={props.onSpellSelect}
@@ -21,10 +22,3 @@ const SpellList = props => {
 };
 
 export default SpellList;
-/*<p className="tc ma1 pa1">Make sure to *Deselect* (at the bottom) the previously selected spells before trying to *Select* another spell or use the filters in the top bar!</p>*/
-// const styles = {
-  
-//   foot: {
-
-//   }
-// };
